Set up structure of game logic
At this point the game is working as intended. Still need to get the
computer choice to print, as currently it just returns. Return seemed
necessary to get the values to the playRound function. Also moved some
variable declarations to global scope, as it is not necessary for most
(or any) of them to be initially declared in functions.